requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
weaviate-client>=4.3.0
python-Levenshtein>=0.21.0
tqdm>=4.65.0
requests>=2.31.0
pyyaml>=6.0.0
jinja2>=3.1.0
tenacity>=8.2.0
psutil>=5.9.0
openai>=1.3.0
joblib>=1.2.0
networkx>=2.8.0
jellyfish>=0.9.0  # For Jaro-Winkler and other string metrics
pytest>=7.3.0  # For testing
jupyter>=1.0.0  # For notebooks
ipykernel>=6.15.0  # For notebooks
pathlib>=1.0.1
pyarrow>=10.0.0  # For efficient file I/O
fastparquet>=0.8.0  # For efficient file I/O
dask>=2023.1.0  # For parallel computing
torch>=2.0.0  # For potential advanced models
prometheus-client>=0.16.0  # For monitoring
loguru>=0.6.0  # Better logging
concurrent-log-handler>=0.9.20  # For robust logging

